+++ bm-common/src/main/java/org/benchmarker/bmcommon/dto/CommonTestResult.java
@@ -45,7 +45,7 @@ public class CommonTestResult {
     @JsonProperty("mttfb_average")
     private String mttfbAverage;
     @JsonProperty("mttfb_percentiles")
-    private Map<Double, Double> MTTFBPercentiles;
+    private Map<Double, Long> MTTFBPercentiles;
     @JsonProperty("tps_average")
     private double tpsAverage;
     @JsonProperty("tps_percentiles")
